Listen on one or more addresses
In order to listen on, for instance, both `::0` and `0.0.0.0`, hokay
needs to be able to listen on more than one address. This change updates
hokay to use `clap` to process command-line arguments and to provide
reasonable help messages.

This change also bumps the version to 0.2.0, as the command-line
interface has changed.